Refactor cli.py into click-depedant cli.py and click-agnostic api.py
