+++ digital_land/cli.py
@@ -1,42 +1,15 @@
 import functools
-import itertools
 import logging
-import os
 import sys
 from collections import defaultdict
 from datetime import date
-from pathlib import Path
 
-import canonicaljson
 import click
 
-from .collect import Collector
-from .collection import Collection, resource_path
-from .convert import Converter
-from .entry_loader import EntryLoader
-from .filter import Filterer
-from .harmonise import Harmoniser
-from .index import Indexer
-from .issues import Issues, IssuesFile
-from .load import LineConverter, load_csv, load_csv_dict
-from .map import Mapper
-from .model.entity import Entity
-from .normalise import Normaliser
-from .organisation import Organisation
-from .pipeline import Pipeline
-from .plugin import get_plugin_manager
-from .repository.entry_repository import EntryRepository
-from .save import save
-from .schema import Schema
-from .lookup import Lookup
-from .slug import Slugger
-from .specification import Specification
-from .transform import Transformer
-from .update import add_source_endpoint, get_failing_endpoints_from_registers
-from .datasette.docker import build_container
+from digital_land.api import DigitalLandApi
 
-PIPELINE = None
-SPECIFICATION = None
+# pyright: reportGeneralTypeIssues=false
+API: DigitalLandApi = None
 
 
 # Custom decorators for common command arguments
@@ -113,20 +86,15 @@ def organisation_path(f):
 @pipeline_dir
 @specification_dir
 def cli(debug, pipeline_name, pipeline_dir, specification_dir):
-    level = logging.DEBUG if debug else logging.INFO
-    logging.basicConfig(level=level, format="%(asctime)s %(levelname)s %(message)s")
-    global PIPELINE
-    global SPECIFICATION
-    PIPELINE = Pipeline(pipeline_dir, pipeline_name)
-    SPECIFICATION = Specification(specification_dir)
+    global API
+    API = DigitalLandApi(debug, pipeline_name, pipeline_dir, specification_dir)
 
 
 @cli.command("fetch")
 @click.argument("url")
 def fetch_cmd(url):
     """fetch a single source endpoint URL, and add it to the collection"""
-    collector = Collector(PIPELINE.name)
-    collector.fetch(url)
+    return API.fetch_cmd(url)
 
 
 @cli.command("collect")
@@ -138,8 +106,7 @@ def fetch_cmd(url):
 @collection_dir
 def collect_cmd(endpoint_path, collection_dir):
     """fetch the sources listed in the endpoint-url column of the ENDPOINT_PATH CSV file"""
-    collector = Collector(PIPELINE.name, Path(collection_dir))
-    collector.collect(endpoint_path)
+    return API.collect_cmd(endpoint_path, collection_dir)
 
 
 #
@@ -152,17 +119,13 @@ def collect_cmd(endpoint_path, collection_dir):
 )
 def index_cmd():
     # TBD: replace with Collection()
-    indexer = Indexer()
-    indexer.index()
+    return API.index_cmd()
 
 
 @cli.command("collection-list-resources", short_help="list resources for a pipeline")
 @collection_dir
 def pipeline_collection_list_resources_cmd(collection_dir):
-    collection = Collection(name=None, directory=collection_dir)
-    collection.load()
-    for resource in sorted(collection.resource.records):
-        print(resource_path(resource, directory=collection_dir))
+    return API.pipeline_collection_list_resources_cmd(collection_dir)
 
 
 @cli.command(
@@ -171,22 +134,13 @@ def pipeline_collection_list_resources_cmd(collection_dir):
 )
 @collection_dir
 def pipeline_collection_pipeline_makerules_cmd(collection_dir):
-    collection = Collection(name=None, directory=collection_dir)
-    collection.load()
-    collection.pipeline_makerules()
+    return API.pipeline_collection_pipeline_makerules_cmd(collection_dir)
 
 
 @cli.command("collection-save-csv", short_help="save collection as CSV package")
 @collection_dir
 def pipeline_collection_save_csv_cmd(collection_dir):
-    try:
-        os.remove(Path(collection_dir) / "log.csv")
-        os.remove(Path(collection_dir) / "resource.csv")
-    except OSError:
-        pass
-    collection = Collection(name=None, directory=collection_dir)
-    collection.load()
-    collection.save_csv()
+    return API.pipeline_collection_save_csv_cmd(collection_dir)
 
 
 #
@@ -195,15 +149,7 @@ def pipeline_collection_save_csv_cmd(collection_dir):
 @cli.command("convert", short_help="convert to a well-formed, UTF-8 encoded CSV file")
 @input_output_path
 def convert_cmd(input_path, output_path):
-    if not output_path:
-        output_path = default_output_path_for("converted", input_path)
-
-    converter = Converter(PIPELINE.conversions())
-    reader = converter.convert(input_path)
-    if not reader:
-        logging.error(f"Unable to convert {input_path}")
-        sys.exit(2)
-    save(reader, output_path)
+    return API.convert_cmd(input_path, output_path)
 
 
 @cli.command("normalise", short_help="removed padding, drop empty rows")
@@ -221,46 +167,19 @@ def convert_cmd(input_path, output_path):
     default=None,
 )
 def normalise_cmd(input_path, output_path, null_path, skip_path):
-    if not output_path:
-        output_path = default_output_path_for("normalised", input_path)
-
-    resource_hash = resource_hash_from(input_path)
-    stream = load_csv(input_path)
-    normaliser = Normaliser(PIPELINE.skip_patterns(resource_hash), null_path=null_path)
-    stream = normaliser.normalise(stream)
-    save(stream, output_path)
+    return API.normalise_cmd(input_path, output_path, null_path, skip_path)
 
 
 @cli.command("map", short_help="map misspelt column names to those in a pipeline")
 @input_output_path
 def map_cmd(input_path, output_path):
-    if not output_path:
-        output_path = default_output_path_for("mapped", input_path)
-
-    resource_hash = resource_hash_from(input_path)
-    fieldnames = intermediary_fieldnames(SPECIFICATION, PIPELINE)
-    mapper = Mapper(
-        fieldnames,
-        PIPELINE.columns(resource_hash),
-        PIPELINE.concatenations(resource_hash),
-    )
-    stream = load_csv_dict(input_path)
-    stream = mapper.map(stream)
-    save(stream, output_path, fieldnames=fieldnames)
+    return API.map_cmd(input_path, output_path)
 
 
 @cli.command("filter", short_help="filter out rows by field values")
 @input_output_path
 def filter_cmd(input_path, output_path):
-    if not output_path:
-        output_path = default_output_path_for("filtered", input_path)
-
-    resource_hash = resource_hash_from(input_path)
-    fieldnames = intermediary_fieldnames(SPECIFICATION, PIPELINE)
-    filterer = Filterer(PIPELINE.filters(resource_hash))
-    stream = load_csv_dict(input_path)
-    stream = filterer.filter(stream)
-    save(stream, output_path, fieldnames=fieldnames)
+    return API.filter_cmd(input_path, output_path)
 
 
 @cli.command(
@@ -271,104 +190,27 @@ def filter_cmd(input_path, output_path):
 @issue_dir
 @organisation_path
 def harmonise_cmd(input_path, output_path, issue_dir, organisation_path):
-    if not output_path:
-        output_path = default_output_path_for("harmonised", input_path)
-
-    resource_hash = resource_hash_from(input_path)
-    issues = Issues()
-
-    collection = Collection()
-    collection.load()
-
-    organisation_uri = Organisation(
-        organisation_path, Path(PIPELINE.path)
-    ).organisation_uri
-    patch = PIPELINE.patches(resource_hash)
-    fieldnames = intermediary_fieldnames(SPECIFICATION, PIPELINE)
-
-    pm = get_plugin_manager()
-
-    harmoniser = Harmoniser(
-        SPECIFICATION,
-        PIPELINE,
-        issues,
-        collection,
-        organisation_uri,
-        patch,
-        pm,
-    )
-
-    stream = load_csv_dict(input_path)
-    stream = harmoniser.harmonise(stream)
-    save(stream, output_path, fieldnames=fieldnames)
-
-    issues_file = IssuesFile(path=os.path.join(issue_dir, resource_hash + ".csv"))
-    issues_file.write_issues(issues)
+    return API.harmonise_cmd(input_path, output_path, issue_dir, organisation_path)
 
 
 @cli.command("transform", short_help="transform")
 @input_output_path
 @organisation_path
 def transform_cmd(input_path, output_path, organisation_path):
-    if not output_path:
-        output_path = default_output_path_for("transformed", input_path)
-
-    organisation = Organisation(organisation_path, Path(PIPELINE.path))
-    schema = SPECIFICATION.pipeline[PIPELINE.name]["schema"]
-    transformer = Transformer(
-        schema,
-        PIPELINE.transformations(),
-        organisation.organisation,
-    )
-    stream = load_csv_dict(input_path)
-    stream = transformer.transform(stream)
-    save(stream, output_path, SPECIFICATION.current_fieldnames(schema))
+    return API.transform_cmd(input_path, output_path, organisation_path)
 
 
 @cli.command("load-entries", short_help="load entries")
 @click.option("--output-path", type=click.Path(), default=None)
 @click.argument("input-paths", nargs=-1, type=click.Path(exists=True))
 def load_entries_cmd(input_paths, output_path):
-    if not output_path:
-        print("missing output path")
-        sys.exit(2)
-
-    repo = EntryRepository(output_path, create=True)
-    loader = EntryLoader(repo)
-
-    total = len(input_paths)
-    for idx, path in enumerate(input_paths, start=1):
-        logging.info("loading %s [%s/%s]", path, idx, total)
-        stream = load_csv_dict(path, include_line_num=True)
-        loader.load(stream)
+    return API.load_entries_cmd(input_paths, output_path)
 
 
 @cli.command("build-dataset", short_help="build dataset")
 @input_output_path
 def build_dataset_cmd(input_path, output_path):
-    repo = EntryRepository(input_path)
-    slugs = repo.list_slugs()
-    logging.info("building dataset with %s slugs", len(slugs))
-    schema = SPECIFICATION.pipeline[PIPELINE.name]["schema"]
-
-    output = filter(
-        lambda x: x["row"],
-        (
-            {
-                "row": Entity(
-                    repo.find_by_slug(slug),
-                    schema,
-                ).snapshot()
-            }
-            for slug in slugs
-        ),
-    )
-
-    save(
-        output,
-        output_path,
-        SPECIFICATION.current_fieldnames(schema),
-    )
+    return API.build_dataset_cmd(input_path, output_path)
 
 
 @cli.command("pipeline", short_help="convert, normalise, map, harmonise, transform")
@@ -392,99 +234,15 @@ def pipeline_cmd(
     organisation_path,
     save_harmonised,
 ):
-    resource_hash = resource_hash_from(input_path)
-    organisation = Organisation(organisation_path, Path(PIPELINE.path))
-    issues = Issues()
-
-    schema = SPECIFICATION.pipeline[PIPELINE.name]["schema"]
-    fieldnames = intermediary_fieldnames(SPECIFICATION, PIPELINE)
-    patch = PIPELINE.patches(resource_hash)
-
-    collection = Collection(name=None, directory=collection_dir)
-    collection.load()
-    line_converter = LineConverter()
-    pm = get_plugin_manager()
-
-    converter = Converter(PIPELINE.conversions())
-
-    normaliser = Normaliser(PIPELINE.skip_patterns(resource_hash), null_path=null_path)
-    mapper = Mapper(
-        fieldnames,
-        PIPELINE.columns(resource_hash),
-        PIPELINE.concatenations(resource_hash),
-    )
-    filterer = Filterer(PIPELINE.filters(resource_hash))
-    harmoniser = Harmoniser(
-        SPECIFICATION,
-        PIPELINE,
-        issues,
-        collection,
-        organisation.organisation_uri,
-        patch,
-        pm,
-    )
-    transformer = Transformer(
-        SPECIFICATION.schema_field[schema],
-        PIPELINE.transformations(),
-        organisation.organisation,
-    )
-
-    key_field = SPECIFICATION.key_field(schema)
-
-    lookup = Lookup(
-        lookups=PIPELINE.lookups(resource_hash),
-        key_field=key_field,
-    )
-
-    slugger = Slugger(
-        SPECIFICATION.pipeline[PIPELINE.name].get("slug-prefix", None),
-        key_field,
-        SPECIFICATION.pipeline[PIPELINE.name].get("scope-field", None),
-    )
-
-    pipeline_funcs = [
-        converter.convert,
-        normaliser.normalise,
-        line_converter.convert,
-        mapper.map,
-        filterer.filter,
-        harmoniser.harmonise,
-    ]
-
-    if save_harmonised:
-        harmonised_path = output_path.replace("transformed", "harmonised")
-        if harmonised_path == output_path:
-            raise ValueError("cannot write harmonised file due to name clash")
-
-        def saver(reader):
-            output_tap, save_tap = itertools.tee(reader)
-            save(
-                save_tap,
-                harmonised_path,
-                fieldnames=intermediary_fieldnames(SPECIFICATION, PIPELINE),
-            )
-            yield from output_tap
-
-        pipeline_funcs.append(saver)
-
-    pipeline_funcs = pipeline_funcs + [
-        transformer.transform,
-        lookup.lookup,
-        slugger.slug,
-    ]
-
-    pipeline = compose(*pipeline_funcs)
-
-    output = pipeline(input_path)
-
-    save(
-        output,
+    return API.pipeline_cmd(
+        input_path,
         output_path,
-        fieldnames=SPECIFICATION.current_fieldnames(schema),
+        collection_dir,
+        null_path,
+        issue_dir,
+        organisation_path,
+        save_harmonised,
     )
-
-    issues_file = IssuesFile(path=os.path.join(issue_dir, resource_hash + ".csv"))
-    issues_file.write_issues(issues)
 
 
 # Endpoint commands
@@ -509,10 +267,9 @@ def pipeline_cmd(
 )
 def collection_check_endpoints_cmd(first_date, log_dir, endpoint_path, last_date):
     """find active endpoints that are failing during collection"""
-    output = get_failing_endpoints_from_registers(
-        log_dir, endpoint_path, first_date.date(), last_date.date()
+    return API.collection_check_endpoints_cmd(
+        first_date, log_dir, endpoint_path, last_date
     )
-    print(canonicaljson.encode_canonical_json(output))
 
 
 @cli.command(
@@ -537,18 +294,9 @@ def collection_add_source_cmd(ctx, collection, endpoint_url, collection_dir):
         str,
         {ctx.args[i]: ctx.args[i + 1] for i in range(0, len(ctx.args), 2)},
     )
-    entry["collection"] = collection
-    entry["endpoint-url"] = endpoint_url
-
-    allowed_names = set(
-        list(Schema("endpoint").fieldnames) + list(Schema("source").fieldnames)
+    return API.collection_add_source_cmd(
+        entry, collection, endpoint_url, collection_dir
     )
-    for key in entry.keys():
-        if key not in allowed_names:
-            logging.error(f"unrecognised argument '{key}'")
-            sys.exit(2)
-
-    add_source_endpoint(entry, directory=collection_dir)
 
 
 @cli.command("build-datasette", short_help="build docker image for datasette")
@@ -557,39 +305,4 @@ def collection_add_source_cmd(ctx, collection, endpoint_url, collection_dir):
 @click.option("--ext", default="sqlite3")
 @click.option("--options", default=None)
 def build_datasette(tag, data_dir, ext, options):
-    datasets = [f"{d}" for d in Path(data_dir).rglob(f"*.{ext}")]
-    for dataset in datasets:
-        if not Path(dataset).exists():
-            print(f"{dataset} not found")
-            sys.exit(1)
-
-    container_id, name = build_container(datasets, tag, options)
-    click.echo("%s dataset successfully packaged" % len(datasets))
-    click.echo(f"container_id: {container_id}")
-    if name:
-        click.echo(f"name: {name}")
-
-
-def resource_hash_from(path):
-    return Path(path).stem
-
-
-def intermediary_fieldnames(specification, pipeline):
-    schema = specification.pipeline[pipeline.name]["schema"]
-    fieldnames = specification.schema_field[schema].copy()
-    replacement_fields = list(pipeline.transformations().keys())
-    for field in replacement_fields:
-        if field in fieldnames:
-            fieldnames.remove(field)
-    return fieldnames
-
-
-def default_output_path_for(command, input_path):
-    return f"var/{command}/{resource_hash_from(input_path)}.csv"
-
-
-def compose(*functions):
-    def compose2(f, g):
-        return lambda x: g(f(x))
-
-    return functools.reduce(compose2, functions, lambda x: x)
+    return API.build_datasette(tag, data_dir, ext, options)
